Fidelity score and LLM fixed
