+++ data/agg_table_allyears_20241118_Nielsen_mascara.xlsx
