+++ data/avg_unit_price_mascara.csv
